batch: Move candidate preview state
operation_candidates.py owned candidate models, materialization, rendering,
and saved preview state. The state-file lifecycle is a persistence concern
rather than candidate construction.

The new operation_candidate_state module owns preview state keys, load/save
logic, and stale preview clearing. Command support code now imports state
helpers from that boundary.

This commit moves candidate preview persistence out of
operation_candidates.py.

Candidate materialization remains focused on merge candidate enumeration,
target previews, fingerprints, and diff rendering.
